+++ core/utils/lume_config.ts
@@ -9,6 +9,7 @@ export const pluginNames = [
   "eta",
   "favicon",
   "feed",
+  "fff",
   "filter_pages",
   "inline",
   "jsx",
@@ -23,6 +24,7 @@ export const pluginNames = [
   "multilanguage",
   "nav",
   "nunjucks",
+  "og_images",
   "on_demand",
   "pagefind",
   "picture",
@@ -30,6 +32,7 @@ export const pluginNames = [
   "prism",
   "pug",
   "reading_info",
+  "redirects",
   "relations",
   "relative_urls",
   "remark",
@@ -50,6 +53,7 @@ export const pluginNames = [
 /** Returns the _config file of a site */
 export async function getConfigFile(
   path?: string,
+  defaultPaths: string[] = ["_config.js", "_config.ts"],
 ): Promise<string | undefined> {
   if (path) {
     try {
@@ -59,9 +63,7 @@ export async function getConfigFile(
     }
   }
 
-  const paths = ["_config.js", "_config.ts"];
-
-  for (const path of paths) {
+  for (const path of defaultPaths) {
     try {
       return await Deno.realPath(path);
     } catch {