+++ plugins/favicon.ts
@@ -31,7 +31,7 @@ export const defaults: Options = {
   favicons: [
     {
       url: "/favicon.ico",
-      size: [16, 32],
+      size: [48],
       rel: "icon",
       format: "ico",
     },
@@ -117,19 +117,20 @@ export default function (userOptions?: Options) {
       for (const page of pages) {
         const document = page.document!;
 
-        if (options.input.endsWith(".svg")) {
-          addIcon(document, {
-            rel: "icon",
-            href: site.url(options.input),
-            type: "image/svg+xml",
-          });
-        }
-
         for (const favicon of options.favicons) {
           addIcon(document, {
             rel: favicon.rel,
             sizes: favicon.size.map((s) => `${s}x${s}`).join(" "),
             href: site.url(favicon.url),
+          });
+        }
+
+        if (options.input.endsWith(".svg")) {
+          addIcon(document, {
+            rel: "icon",
+            sizes: "any",
+            href: site.url(options.input),
+            type: "image/svg+xml",
           });
         }
       }