+++ tests/og_images.test.ts
@@ -1,0 +1,15 @@
+import { assertSiteSnapshot, build, getSite } from "./utils.ts";
+import ogImages from "../plugins/og_images.ts";
+
+Deno.test("OpenGraph images", async (t) => {
+  const site = getSite({
+    src: "og_images",
+  });
+
+  site.use(ogImages({
+    cache: false,
+  }));
+
+  await build(site);
+  await assertSiteSnapshot(t, site);
+});