+++ tests/sitemap.test.ts
@@ -1,6 +1,7 @@
 import { assertSiteSnapshot, build, getSite } from "./utils.ts";
 import sitemap from "../plugins/sitemap.ts";
 import multilanguage from "../plugins/multilanguage.ts";
+import filterPages from "../plugins/filter_pages.ts";
 
 Deno.test("Sitemap plugin", async (t) => {
   const site = getSite({
@@ -27,6 +28,22 @@ Deno.test("Sitemap plugin with a multilanguage plugin", async (t) => {
   }));
   site.use(sitemap({
     lastmod: "",
+  }));
+
+  await build(site);
+  await assertSiteSnapshot(t, site);
+});
+
+Deno.test("Sitemap plugin with filter_pages plugin", async (t) => {
+  const site = getSite({
+    src: "normal",
+    location: new URL("https://example.com/"),
+  });
+
+  site.use(sitemap());
+  site.ignore("static.yml");
+  site.use(filterPages({
+    fn: (page) => page.data.url !== "/pages/new-name/page7/",
   }));
 
   await build(site);